Добавлены механизмы-ловушки, заменена логика отображения ранения, добалвена новая карта, добавлен переход на след карту

+++ Game/DarkCastle/DarkCastle/Headers/bonus_logic.h
@@ -9,7 +9,7 @@ struct BonusLogic {
 	float current_distance;
 	float speed;
 	int value;
-	int duration;
+	float duration;
 	bool picked_up;
 	sf::FloatRect* rect;
 	BonusType bonus_type;
